+++ src/features/speedrun/season3/callbacks/postNewRoom.ts
@@ -1,38 +1,33 @@
 import {
-  EntityFlag,
-  EntityType,
   GameStateFlag,
+  GridEntityType,
   LevelStage,
   RoomType,
+  TeleporterState,
 } from "isaac-typescript-definitions";
 import {
   game,
   getBlueWombDoor,
-  getRoomGridIndex,
+  inStartingRoom,
   isRoomInsideGrid,
-  log,
   onRepentanceStage,
-  removeAllCollectibles,
-  removeAllRocks,
   removeDoor,
-  spawnNPC,
   spawnTeleporter,
-  VectorZero,
 } from "isaacscript-common";
 import { ChallengeCustom } from "../../../../enums/ChallengeCustom";
-import { EntityTypeCustom } from "../../../../enums/EntityTypeCustom";
 import g from "../../../../globals";
 import { getNumRoomsEntered } from "../../../utils/numRoomsEntered";
 import { isOnFirstCharacter } from "../../speedrun";
 import { resetSeason3StartingRoomSprites } from "../startingRoomSprites";
-import v, {
+import {
   season3HasDogmaGoal,
   season3HasHushGoal,
   season3HasMegaSatanGoal,
 } from "../v";
 
 const LEFT_OF_CENTER_GRID_INDEX = 65;
-const TOP_CENTER_GRID_INDEX = 22;
+
+const LEFT_OF_TOP_DOOR_GRID_INDEX = 20;
 
 export function season3PostNewRoom(): void {
   const challenge = Isaac.GetChallenge();
@@ -46,26 +41,34 @@ export function season3PostNewRoom(): void {
     resetSeason3StartingRoomSprites();
   }
 
-  checkSpawnMegaSatanTeleporter();
+  checkMegaSatanTeleporter();
   checkDadsNoteRoom();
   checkBlueWombRoom();
 }
 
-function checkSpawnMegaSatanTeleporter() {
+function checkMegaSatanTeleporter() {
   const stage = g.l.GetStage();
   const isFirstVisit = g.r.IsFirstVisit();
-  const startingRoomGridIndex = g.l.GetStartingRoomIndex();
-  const roomGridIndex = getRoomGridIndex();
 
-  if (
-    stage === LevelStage.DARK_ROOM_CHEST &&
-    roomGridIndex === startingRoomGridIndex &&
-    isFirstVisit &&
-    season3HasMegaSatanGoal() &&
-    !isOnFirstCharacter()
-  ) {
-    spawnTeleporter(TOP_CENTER_GRID_INDEX);
-    v.room.megaSatanTeleporterSpawned = true;
+  if (stage !== LevelStage.DARK_ROOM_CHEST || !inStartingRoom()) {
+    return;
+  }
+
+  // Handle spawning the teleporter.
+  if (isFirstVisit && season3HasMegaSatanGoal() && !isOnFirstCharacter()) {
+    // We don't want to spawn the teleporter exactly where the top door would be because that is the
+    // position that they will return to if they e.g. die in the Mega Satan fight with Dead Cat.
+    spawnTeleporter(LEFT_OF_TOP_DOOR_GRID_INDEX);
+  }
+
+  // If we return to the starting room after the teleporter has already been activated, then we have
+  // to manually change the state to allow the player to return to the Mega Satan room.
+  const gridEntity = g.r.GetGridEntity(LEFT_OF_TOP_DOOR_GRID_INDEX);
+  if (gridEntity !== undefined) {
+    const gridEntityType = gridEntity.GetType();
+    if (gridEntityType === GridEntityType.TELEPORTER) {
+      gridEntity.State = TeleporterState.NORMAL;
+    }
   }
 }
 
@@ -79,7 +82,6 @@ function checkDadsNoteRoom() {
   const roomInsideGrid = isRoomInsideGrid();
 
   if (
-    !v.run.killedDogma &&
     season3HasDogmaGoal() &&
     stage === LevelStage.DEPTHS_2 &&
     repentanceStage &&
@@ -87,6 +89,7 @@ function checkDadsNoteRoom() {
     roomInsideGrid &&
     backwardsPathInit
   ) {
+    /*
     removeAllCollectibles(); // Remove Dad's Note.
 
     // We can't spawn Dogma in the center of the room, because then it will overlap with the TV and
@@ -99,6 +102,8 @@ function checkDadsNoteRoom() {
     // The Dad's Note room will always have rocks in each of the four corners. Their presence makes
     // the fight a lot harder, so remove them.
     removeAllRocks();
+    */
+    // Take them directly to Home to avoid wasting time.
   }
 }
 
@@ -109,6 +114,7 @@ function checkDadsNoteRoom() {
  *
  * Using a room clear delay effect for this purpose does not work.
  */
+/*
 function spawnRoomClearDelayNPC() {
   const roomClearDelayNPC = spawnNPC(
     EntityTypeCustom.ROOM_CLEAR_DELAY_NPC,
@@ -119,6 +125,7 @@ function spawnRoomClearDelayNPC() {
   roomClearDelayNPC.ClearEntityFlags(EntityFlag.APPEAR);
   log('Spawned the "Room Clear Delay NPC" custom entity (for Dogma).');
 }
+*/
 
 /** Guard against the player accidentally going to Hush. */
 function checkBlueWombRoom() {