+++ docs/history.md
@@ -11,6 +11,7 @@
 - In season 3, the static fade out effect after killing Dogma is now prevented. (Thanks makia)
 - Fixed the bug where defeating a wave in the Boss Rush did not grant any charges. (Thanks 3kch1)
 - Fixed the bug where the photos could spawn in the Dad's Note room under certain conditions. (Thanks makia)
+- Fixed the bug where the Mega Satan teleporter would only work one time.
 
 ### _v0.84.0_ - August 22nd, 2022